requirements.txt
asgiref==3.7.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.2.7
idna==3.4
pika==1.3.1
requests==2.28.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.13
websockets==10.4
djangorestframework==3.14.0
google-api-core==2.13.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-secret-manager==2.16.4
django-storages[google]==1.13.2
gunicorn==20.1.0
psycopg2-binary
django-environ==0.10.0
pytest==7.4.3
pytest-django==4.6.0
